03_software engineering/Approach 1/workspace/smart-fridge-app/backend/requirements.txt
flask
flask-restx
mysql-connector-python
flask-cors
flask-login
python-dotenv
